#328: rename filter and refactor code
