+++ .github/workflows/deploy_server_dev.yml
@@ -85,8 +85,8 @@ jobs:
         ENABLE_LOCAL_LOGIN: ${{ secrets.ENABLE_LOCAL_LOGIN_DEV }}
         WFH_DEFINITIONS_ID: ${{ secrets.WFH_DEFINITIONS_ID_DEV }}
         URL_WEB: ${{ secrets.URL_WEB_DEV }}
-        MEZON_SECRET_KEY_HEADER_NAME: ${{ secrets.MEZON_SECRET_KEY_HEADER_NAME_DEV }}
-        MEZON_X_SECRET_KEY: ${{ secrets.MEZON_X_SECRET_KEY_DEV }}
+        API_SECRET_KEY_HEADER_NAME: ${{ secrets.API_SECRET_KEY_HEADER_NAME_DEV }}
+        API_X_SECRET_KEY: ${{ secrets.API_X_SECRET_KEY_DEV }}
 
     - name: Migrate Database
       run: |
