+++ appsettings_exp.json
@@ -63,8 +63,8 @@
     "ApiUrl": "KOMU_API_URL",
     "XSecretKey": "KOMU_X_SK"
   },
-  "MezonConfiguration": {
-    "XSecretKey": "MEZON_SECRET_KEY_HEADER_NAME",
-    "SecretKeyHeaderName": "MEZON_X_SECRET_KEY"
+  "ApiConfiguration": {
+    "XSecretKey": "API_SECRET_KEY_HEADER_NAME",
+    "SecretKeyHeaderName": "API_X_SECRET_KEY"
   }
 }
