+++ src/W2.Application/Utils/ApiKeyAuth.cs
@@ -1,25 +1,21 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Filters;
+using Microsoft.Extensions.DependencyInjection;
 using System.Linq;
 using Microsoft.Extensions.Options;
 
-namespace W2.Mezon;
+namespace W2.Utils;
 
-public class SecretKeyFilterAttribute : ActionFilterAttribute
+public class ApiKeyAuth : ActionFilterAttribute
 {
-    private readonly Configurations.MezonConfiguration _mezonConfiguration;
-
-    public SecretKeyFilterAttribute(
-        IOptions<Configurations.MezonConfiguration> mezonConfigurationOptions
-    )
-    {
-        _mezonConfiguration = mezonConfigurationOptions.Value;
-    }
-
     public override void OnActionExecuting(ActionExecutingContext context)
     {
-        var secretKeyHeaderName = _mezonConfiguration.SecretKeyHeaderName;
-        var secretKey = _mezonConfiguration.XSecretKey;
+        var serviceProvider = context.HttpContext.RequestServices;
+
+        var apiConfiguration = serviceProvider.GetRequiredService<IOptions<Configurations.ApiConfiguration>>().Value;
+
+        var secretKeyHeaderName = apiConfiguration.SecretKeyHeaderName;
+        var secretKey = apiConfiguration.XSecretKey;
         if (!context.HttpContext.Request.Headers.TryGetValue(secretKeyHeaderName, out var extractedSecretKey))
         {
             context.Result = new UnauthorizedObjectResult(new { Message = "Invalid or Missing API Key!" });
