+++ src/W2.Domain/Configurations/ApiConfiguration.cs
@@ -1,6 +1,6 @@
 ﻿namespace W2.Configurations;
 
-public class MezonConfiguration
+public class ApiConfiguration
 {
     public string SecretKeyHeaderName { get; set; }
     public string XSecretKey { get; set; }
