+++ src/W2.Web/appsettings.json
@@ -52,7 +52,7 @@
     "Ant": "https://ant.ncc.asia",
     "HrmCode": "dont_use_for_now"
   },
-  "MezonConfiguration": {
+  "ApiConfiguration": {
     "XSecretKey": "your-secret-key",
     "SecretKeyHeaderName": "x-secret-key"
   }
